moved MainMongoDB outside Use Case layer to gateways folder, and refactored code as necessary
