--- src/main/java/gateways/BudgetingGateway.java
@@ -1,4 +1,0 @@
-package gateways;
-
-public interface BudgetingGateway {
-}
