--- src/main/java/gateways/FoodSuggestionsGateway.java
@@ -1,4 +1,0 @@
-package gateways;
-
-public interface FoodSuggestionsGateway {
-}
