--- src/main/java/gateways/MakeOrderGateway.java
@@ -1,4 +1,0 @@
-package gateways;
-
-public class MakeOrderGateway {
-}
