--- src/main/java/gateways/PriceFilteringGateway.java
@@ -1,4 +1,0 @@
-package gateways;
-
-public interface PriceFilteringGateway {
-}
