--- src/main/java/gateways/RatingsGateway.java
@@ -1,4 +1,0 @@
-package gateways;
-
-public interface RatingsGateway {
-}
