--- src/main/java/gateways/RestaurantFilteringGateway.java
@@ -1,26 +1,0 @@
-package gateways;
-import usecases.MainMongoDB;
-
-import java.util.ArrayList;
-
-public class RestaurantFilteringGateway {
-//    public static ArrayList<String> getListOfRestaurants() {
-//        return MainMongoDB.getAllRestaurants();
-//    }
-//
-//    public static String getRestaurantCuisine(String name) {
-//        return MainMongoDB.getRestaurantAttribute(name, "cuisine");
-//    }
-//
-//    public static String getRestaurantPriceRange(String name) {
-//        return MainMongoDB.getRestaurantAttribute(name, "priceRange");
-//    }
-//
-//    public static String getRestaurantFoodType(String name) {
-//        return MainMongoDB.getRestaurantAttribute(name, "foodType");
-//    }
-//
-//    public static String getRestaurantAvgRating(String name) {
-//        return MainMongoDB.getRestaurantAttribute(name, "avgRating");
-//    }
-}
