+++ src/main/java/usecases/login/Login.java
@@ -1,8 +1,7 @@
 //Swami Shriji
 package usecases.login;
-import com.sun.tools.javac.Main;
 import usecases.LoginDAI;
-import usecases.MainMongoDB;
+import gateways.MainMongoDB;
 
 public class Login {
     private final LoginDAI loginDAI = new MainMongoDB();
